+++ src/types/context.ts
@@ -12,6 +12,8 @@ import WebSocket from "./webSocket"
 import { EventHandlerMap } from "./event"
 import { HttpRequest, WsClose, WsConnect, WsMessage } from "./external"
 import Reference, { RefListener } from "../classes/reference"
+import { Content } from "../functions/parseContent"
+import DataStat from "../classes/dataStat"
 
 export type Hours =
 	| '0' | '1' | '2' | '3' | '4'
@@ -20,13 +22,36 @@ export type Hours =
 	| '15' | '16' | '17' | '18' | '19'
 	| '20' | '21' | '22' | '23'
 
+type ExecuteFound = {
+	event: 'none' | keyof EventHandlerMap<any, any>
+	found: true
+	route: HTTP | WebSocket
+	file: null
+}
+
+type ExecuteFoundStatic = {
+	event: 'none' | keyof EventHandlerMap<any, any>
+	found: true
+	route: Static
+	file: string
+}
+
+type ExecuteNotFound = {
+	event: 'none' | keyof EventHandlerMap<any, any>
+	found: false
+	route: null
+	file: null
+}
+
+type Execute = ExecuteFound | ExecuteFoundStatic | ExecuteNotFound
+
 export type InternalEvents = {
 	startRequest(): void
 	requestAborted(): void
 }
 
 export type LocalContext = {
-	/** The Current Queue for Async-Sync Functions */ queue: Task[]
+	/** The Code to execute custom */ executeSelf: () => boolean | Promise<boolean>
 	/** The Previous Hours as an Array */ previousHours: Hours[]
 	/** The Parsed Request URL */ url: URLObject
 	/** Whether to Continue ending the Request */ continueSend: boolean
@@ -41,9 +66,9 @@ export type LocalContext = {
 	/** An Event Emitter Responsible for all Events */ events: TypedEventEmitter<InternalEvents>
 	/** A Boolean that keeps track whether the Request is Aborted */ isAborted: boolean
 	/** The Reference Listeners to keep track of and delete */ refListeners: { ref: Reference, refListener: RefListener }[]
+
 	/** The Function to handle an Error in an Async Scenario */ handleError(err: unknown): void
-	/** Schedule an Async Task for Execution */ scheduleQueue(type: Task['type'], callback: Task['function']): void
-	/** Run all current Functions contained in the Queue */ runQueue(): Promise<Error | null>
+	/** Set the Code to run when executeCode is false */ setExecuteSelf(callback: LocalContext['executeSelf']): void
 
 	/** The Current Request Body */ body: {
 		/** The Body as Chunks */ chunks: Buffer[]
@@ -51,19 +76,16 @@ export type LocalContext = {
 		/** The Body as a parsed Object */ parsed: any
 	}
 
-	/** The Execute Object */ execute: {
-		/** The Route Object that was found */ route: HTTP | Static | WebSocket | null
-		/** The File to Read when Route is Static */ file: string | null
-		/** Whether the Route exists */ exists: boolean
-		/** The Event to execute instead of the Route */ event: 'none' | keyof EventHandlerMap<any, any>
-	}
+	/** The Execute Object */ execute: Execute
 
 	/** The Response Object */ response: {
-		/** The Headers to Respond with */ headers: Record<string, Buffer | string>
+		/** The Headers to Respond with */ headers: Record<string, Content>
 		/** The HTTP Status to Respond with */ status: number
 		/** The HTTP Status Message to concat to the Code */ statusMessage: string | undefined
 		/** Whether the Current Content is Compressed */ isCompressed: boolean
-		/** The Raw Content to Send */ content: Buffer
+
+		/** The Raw Content to Send */ content: Content
+		/** Whether to prettify Content */ contentPrettify: boolean
 	}
 }
 
@@ -72,14 +94,14 @@ export type GlobalContext = {
 	/** The File -> Content Type Mapping */ contentTypes: Record<string, string>
 	/** The Default HTTP Headers List */ defaultHeaders: Record<string, Buffer>
 	/** The HTTP Server Options */ options: DeepRequired<Options>
-	/** The Request Count */ requests: Record<Hours | 'total', number>
+	/** The Request Count */ requests: DataStat
 	/** The Middlewares to run */ middlewares: MiddlewareInitted[]
 
 	/** The WebSocket Stats */ webSockets: {
-		/** The Amount of Sockets Opened */ opened: Record<Hours | 'total', number>
+		/** The Amount of Sockets Opened */ opened: DataStat
 		/** The Amount of Socket Messages recieved */ messages: {
-			/** The Incoming Message Count */ incoming: Record<Hours | 'total', number>
-			/** The Outgoing Message Count */ outgoing: Record<Hours | 'total', number>
+			/** The Incoming Message Count */ incoming: DataStat
+			/** The Outgoing Message Count */ outgoing: DataStat
 		}
 	}
 
@@ -91,8 +113,8 @@ export type GlobalContext = {
 	}
 
 	/** The Data Stats */ data: {
-		/** The Incoming Data Count */ incoming: Record<Hours | 'total', number>
-		/** The Outgoing Data Count */ outgoing: Record<Hours | 'total', number>
+		/** The Incoming Data Count */ incoming: DataStat
+		/** The Outgoing Data Count */ outgoing: DataStat
 	}
 
   /** The Routes */ routes: {